Moved Hashing from Core to Plugins
