+++ Plugins/Hashing/include/Acts/Plugins/Hashing/HashingAlgorithm.ipp
@@ -6,7 +6,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-#include "Acts/Seeding/Hashing/HashingAnnoy.hpp"
+#include "Acts/Plugins/Hashing/HashingAnnoy.hpp"
 #include "Acts/Seeding/detail/UtilityFunctions.hpp"
 
 #include <memory>
